Extract AC::Parameters equality tests to own file
